+++ test/io/client_test.dart
@@ -8,7 +8,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:http/http.dart' as http;
-import 'package:http/src/io_client.dart' as http_io;
+import 'package:http/io_client.dart' as http_io;
 import 'package:test/test.dart';
 
 import 'utils.dart';
@@ -120,5 +120,16 @@ void main() {
     var headers = jsonDecode(bytesString)['headers'] as Map<String, dynamic>;
     var contentType = (headers['content-type'] as List).single;
     expect(contentType, startsWith('multipart/form-data; boundary='));
+  });
+
+  test('detachSocket returns a socket from an IOStreamedResponse', () async {
+    var ioClient = HttpClient();
+    var client = http_io.IOClient(ioClient);
+    var request = http.Request('GET', serverUrl);
+
+    var response = await client.send(request);
+    var socket = await response.detachSocket();
+
+    expect(socket, isNotNull);
   });
 }